Fix bug on Room Creation:
We were creating a new WebSocket instance everytime a new room was created, so it was leading to multiple websocket servers being created for different rooms,
and all of them were trying to listen to the same http server, which was causing the already connected sockets to hang-up and the whole api to stop.

So now, we are creating a noServer WebSockets, and filtering the upgrade connections to direct the requests to the required WebSocket.